# engine/src/main/resources/swagger/cromwell.yaml
...
swagger: '2.0'
...
info:
  title: Cromwell Server REST API
  description: Describes the REST API provided by a Cromwell server
  license:
    name: BSD
    url: 'https://github.com/broadinstitute/cromwell/blob/develop/LICENSE.txt'
  version: '30'
produces:
  - application/json
paths:
...
  '/api/workflows/{version}/{id}/releaseHold':
    post:
      operationId: releaseHold
      summary: Switch a workflow from 'On Hold' to 'Submitted' status
      description: Request Cromwell to release the hold on a workflow. It will switch the status of a workflow from 'On Hold' to 'Submitted' so it can be picked for running. For instance this might be necessary in cases where you have submitted a workflow with workflowOnHold = true.
      parameters:
        - $ref: '#/parameters/versionParam'
        - $ref: '#/parameters/singleId'
      tags:
        - Workflows
      responses:
        '200':
          $ref: '#/responses/OkIdAndStatus'
        '400':
          $ref: '#/responses/BadRequest'
        '403':
          $ref: '#/responses/Forbidden'
        '404':
          $ref: '#/responses/NotFound'
        '500':
          $ref: '#/responses/ServerError'
...
  '/api/workflows/{version}/{id}/metadata':
    get:
      operationId: metadata
      summary: Get workflow and call-level metadata for a specified workflow
      parameters:
        - $ref: '#/parameters/versionParam'
        - $ref: '#/parameters/singleId'
        - name: includeKey
          description: >
            When specified, filters metadata to only return fields with names which begins with this value.
            This key is used relative to the root of the response *and* relative to each call's metadata fields.
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
          in: query
        - name: excludeKey
          description: >
            When specified, filters metadata to not return any field with a name which begins with this value.
            This key is used relative to the root of the response *and* relative to each call's metadata fields.
            Use 'calls' to filter out all call level metadata.
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
          in: query
        - name: expandSubWorkflows
          description: >
            When true, metadata for sub workflows will be fetched and inserted automatically in the metadata response.
          required: false
          type: boolean
          in: query
      tags:
        - Workflows
      responses:
        '200':
          description: Successful Request
          schema:
            $ref: '#/definitions/WorkflowMetadataResponse'
        '400':
          $ref: '#/responses/BadRequest'
        '404':
          $ref: '#/responses/NotFound'
        '500':
          $ref: '#/responses/ServerError'
...
  '/api/workflows/{version}/backends':
    get:
      operationId: backends
      summary: List the supported backends
      description: Returns the backends supported by this Cromwell server, as well as the default backend.
      parameters:
        - $ref: '#/parameters/versionParam'
      tags:
        - Workflows
      responses:
        '200':
          description: Successful Request
          schema:
            $ref: '#/definitions/BackendResponse'
...
responses:
  OkIdAndStatus:
    description: Successful Request
    schema:
      $ref: '#/definitions/WorkflowIdAndStatus'
...
  BadRequest:
    description: Malformed Workflow ID
  Forbidden:
    description: Workflow in terminal status
  NotFound:
    description: Workflow ID Not Found
  ServerError:
    description: Internal Error
parameters:
  singleId:
    name: id
    description: A workflow ID
    required: true
    type: string
    in: path
    pattern: ^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$
...
  versionParam:
    name: version
    description: Cromwell API Version
    required: true
    type: string
    in: path
    default: v1 # This really should be an enum but swagger-ui seems to not like it
definitions:
...
  WorkflowIdAndStatus:
    required:
      - id
      - status
    properties:
      id:
        type: string
        description: The identifier of the workflow
        example: 00001111-2222-3333-aaaa-bbbbccccdddd
      status:
        type: string
        description: The status of the workflow
        example: Submitted
  WorkflowMetadataResponse:
    description: Workflow and call level metadata
    required:
      - id
      - status
      - submission
    properties:
      id:
        type: string
        description: The identifier of the workflow
      status:
        type: string
        description: The status of the workflow
      submission:
        type: string
        format: date-time
        description: Submission datetime of the workflow in ISO8601 format with milliseconds
      start:
        type: string
        format: date-time
        description: Start datetime of the workflow in ISO8601 format with milliseconds
      end:
        type: string
        format: date-time
        description: End datetime of the workflow in ISO8601 format with milliseconds
      inputs:
        type: object
        description: Map of input keys to input values
      outputs:
        type: object
        description: Map of output keys to output values
      calls:
        $ref: '#/definitions/CallMetadata'
      failures:
        $ref: '#/definitions/FailureMessage'
  CallMetadata:
    description: Call level metadata
    required:
      - inputs
      - executionStatus
    properties:
      inputs:
        type: object
        description: Mapping of input fully qualified names to stringified values
      executionStatus:
        type: string
        description: Status in Cromwell execution terms.
      backend:
        type: string
        description: The type of backend on which the call executed (e.g. JES, SGE, Local)
      backendStatus:
        type: string
        description: Status in backend-specific terms.  Currently this will only be defined for the JES backend.
      start:
        type: string
        format: date-time
        description: Start datetime of the call execution in ISO8601 format with milliseconds
      end:
        type: string
        format: date-time
        description: End datetime of the call execution in ISO8601 format with milliseconds
      jobId:
        type: string
        description: Backend-specific job ID
      failures:
        $ref: '#/definitions/FailureMessage'
      returnCode:
        type: integer
        description: Call execution return code
      stdout:
        type: string
        description: Path to the standard output file for this call
      stderr:
        type: string
        description: Path to the standard error file for this call
      backendLogs:
        type: object
        description: Paths to backend specific logs for this call
  FailureMessage:
    description: Failure messages
    required:
      - failure
      - timestamp
    properties:
      failure:
        type: string
        description: The failure message
      timestamp:
        type: string
        format: date-time
        description: The time at which this failure occurred
...
  BackendResponse:
    required:
      - supportedBackends
      - defaultBackend
    properties:
      supportedBackends:
        type: array
        description: The backends supported by this server
        items:
          type: string
      defaultBackend:
        type: string
        description: The default backend of this server
...